add upgraded setuptools to dockerfile
A change in the installations of the upstream connexion project in
conjunction with an older version of setuptools in the openshift-spark
image is causing build issues for the ophicleide-trainer. This patch
will make sure that setuptools gets upgraded, it is being added in the
Dockerfile as it is needs to be in place before the pip installation for
the requirements.