add driver host command to spark-submit
This change adds an option to pass a driver hostname to the spark-submit
command. It is being added to help work around issues in spark 2.3+
wherein the workers needs an fqdn back to the driver application. By
specifying the `DRIVER_HOST` environment variable, the application will
inject a configuration command to the spark-submit.